refactor: map 관련 컴포넌트 그룹화
