+++ src/components/domains/products/map/MapPoint.tsx
